Add support for scalars and numpy in torch.testing
Add support to compare scalars as well as `np.ndarray`'s with torch.testing. We are reusing the mathcing functionality that is already in place for tensors, by casting the inputs. The approach can easily extended if we want to support other input types as long as they can be cast to a tensor.

ghstack-source-id: 2445fcb14d53200335607601fd2210c0b8603ddd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55786